@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <AssemblyName>Tiled2Dmap.CLI</AssemblyName>
     <RootNamespace>Tiled2Dmap.CLI</RootNamespace>
   </PropertyGroup>
@@ -29,7 +29,7 @@
   <ItemGroup>
     <PackageReference Include="BCnEncoder.Net" Version="2.1.0" />
     <PackageReference Include="BCnEncoder.Net.ImageSharp" Version="1.1.1" />
-    <PackageReference Include="Cocona" Version="2.0.2" />
+    <PackageReference Include="Cocona" Version="2.2.0" />
     <PackageReference Include="Microsoft.Extensions.FileProviders.Embedded" Version="7.0.2" />
     <PackageReference Include="MonoGame.Framework.DesktopGL" Version="3.8.0.1641" />
     <PackageReference Include="SixLabors.ImageSharp" Version="2.1.3" />
